reviews_service/requirements.txt
fastapi==0.115.12
pydantic==2.11.1
uvicorn==0.20.0
pydantic-settings==2.8.1
motor==3.7.0
pymongo==4.12.0
python-multipart==0.0.20
aiokafka==0.12.0
kafka-python==2.2.4
passlib[bcrypt]
python-jose[cryptography]
prometheus_client==0.21.1
loguru==0.7.3